+++ src/components/browse/dataset/datasetHeader.tsx
@@ -6,7 +6,7 @@ import { useNavigate } from "react-router-dom";
 import { handleFilterAndSearch } from "../../../utils/utils";
 import { querySearchService } from "../../../api/browse";
 
-interface dataSetListHeaderProps {
+interface datasetHeaderProps {
   dsCount: number;
   searchParams: URLSearchParams;
   facets: facetModel[] | null;
@@ -23,7 +23,7 @@ interface dataSetListHeaderProps {
   setPage: Dispatch<SetStateAction<number>>;
 }
 
-const DatasetListHeader = (props: dataSetListHeaderProps) => {
+const DatasetHeader = (props: datasetHeaderProps) => {
   const navigate = useNavigate();
 
   const [searchResults, setSearchResults] =
@@ -98,7 +98,7 @@ const DatasetListHeader = (props: dataSetListHeaderProps) => {
       <Col lg={7} md={7} sm={7} xl={7} xs={7} xxl={7} className="ps-3 offset-3">
         <div className="ps-3 pe-0">
           {props.searchParams.get("q") !== undefined &&
-          props.searchParams.get("q") !== null ? (
+            props.searchParams.get("q") !== null ? (
             <Badge
               key={props.searchParams.get("q")}
               className="py-1 m-0 me-2 overflow-hidden fs-9 bg-white text-black border border-secondary fw-normal"
@@ -161,7 +161,7 @@ const DatasetListHeader = (props: dataSetListHeaderProps) => {
       <Col lg={2} md={2} sm={2} xl={2} xs={2} xxl={2} className="text-end pe-4">
         <Badge className="py-2 px-2 bg-primary me-1">
           {props.searchParams.get("f") !== undefined &&
-          props.searchParams.get("f") !== null
+            props.searchParams.get("f") !== null
             ? "Datasets Found:"
             : "Total Datasets:"}
           &nbsp;{props.dsCount !== -1 ? props.dsCount : 0}
@@ -171,4 +171,4 @@ const DatasetListHeader = (props: dataSetListHeaderProps) => {
   );
 };
 
-export default DatasetListHeader;
+export default DatasetHeader;
