+++ src/components/browse/dataset/datasetPagination/datasetPagination.tsx
@@ -1,10 +1,9 @@
 import React, { Dispatch, SetStateAction } from "react";
 import ReactPaginate from "react-paginate";
-import { searchResponseModel } from "../../../models/dataset";
-import { facetFilterModel } from "../../../models/facets";
-import { handleFilterAndSearch } from "../../../utils/utils";
 import { useNavigate } from "react-router-dom";
-import { scrollUp } from "../../../utils/utils";
+import { scrollUp, handleFilterAndSearch } from "../../../../utils/utils";
+import { searchResponseModel } from "../../../../models/dataset";
+import { facetFilterModel } from "../../../../models/facets";
 
 interface dataSetPaginationProps {
   dsCount: number;
